python-flask/src/app/requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
redis==4.4.2
Werkzeug==2.2.2
azure-cosmos==4.3.0

jsons~=1.6.3
python-dotenv~=0.21.1
pytest~=7.2.1